+++ src/main/java/com/programmers/app/timer/Timer.java
@@ -8,15 +8,15 @@ import com.programmers.app.timer.dto.TimerJson;
 
 public class Timer {
 
-    private long bookNumber;
+    private int bookNumber;
     private LocalDateTime arrangementBegunAt;
 
-    public Timer (long bookNumber, LocalDateTime arrangementBegunAt) {
+    public Timer (int bookNumber, LocalDateTime arrangementBegunAt) {
         this.bookNumber = bookNumber;
         this.arrangementBegunAt = arrangementBegunAt;
     }
 
-    public long getBookNumber() {
+    public int getBookNumber() {
         return bookNumber;
     }
 
@@ -25,7 +25,7 @@ public class Timer {
         return completedAt.isBefore(localDateTime);
     }
 
-    public boolean isMatching(long bookNumber) {
+    public boolean isMatching(int bookNumber) {
         return this.bookNumber == bookNumber;
     }
